% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estaciones_unidas}
\alias{estaciones_unidas}
\title{Datos meteorológicos unificados de 5 estaciones.}
\format{
Un tibble con varias filas y 35 variables (o las que correspondan).
}
\source{
Script de procesamiento interno del paquete.
}
\usage{
estaciones_unidas
}
\description{
Un data frame (tibble) que contiene los datos combinados de las
estaciones NH0046, NH0098, NH0437, NH0472 y NH0910.
}
\keyword{datasets}
