% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Grafica el promedio mensual de temperatura}
\usage{
grafico_temperatura_mensual(datos, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{datos}{Un data frame con datos de estaciones (ej: 'estaciones_unidas').
Debe tener 'fecha', 'id' y 'temperatura_abrigo_150cm'.}

\item{colores}{Un vector de nombres de colores. Si es NULL (por defecto),
se seleccionarán colores aleatorios.}

\item{titulo}{El título del gráfico. El valor por defecto es "Temperatura".}
}
\value{
Un objeto ggplot.
}
\description{
Genera un gráfico de líneas de la temperatura media mensual
para una o más estaciones.
}
